refactored to be more consistent with other features
